kivy_frontend/backup/requirements.txt
appdirs==1.4.4
buildozer==1.4.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==2.1.1
colorama==0.4.5
Cython==0.29.32
distlib==0.3.6
docutils==0.19
filelock==3.8.0
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
Kivy-Garden==0.1.5
kivymd==0.104.2
MarkupSafe==2.1.1
pep517==0.6.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
ptyprocess==0.7.0
Pygments==2.13.0
pytoml==0.1.21
requests==2.28.1
sh==1.14.3
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.5
zipp==3.8.1
